refactor: rename folder from Services to useServices
